Feature branch + additions from pull req description

+++ algorithm/sort/includes/InsertionSort.hpp
@@ -1,9 +1,9 @@
 #ifndef INSERTIONSORT_HPP
 #define INSERTIONSORT_HPP
-#include <cstdint>
+#include <cstddef>
 #include <vector>
 
-void InsertionSort(std::vector <uint64_t > & arr);
+void InsertionSort(std::vector <size_t > & arr);
 
 #endif /* INSERTIONSORT_HPP */
